requirements.txt
pillow
numpy
opencv-python
torch
diffusers
transformers
accelerate
tqdm
exifread